move appcreator to a separate file
